backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
pdfplumber==0.10.4
pydantic==2.6.1
python-dotenv==1.0.1
beautifulsoup4==4.12.3
